requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch         # comment out if not using PyTorch
numpy

